recover-broken-vdb: ask file(1) if file is shared object
We should've used the logic as the detector
script (which does indeed call file(1)), but
we're doing it now at least.

This avoids false positives on e.g. pkgcore's
dolib (which is actually a script).

Avoiding complicating the code by minimising
file(1) calls for now given how it's structured,
but may need to do that. If necessary, we can
call it once-per-package with a list of
candidate shared objects.

Reported-by: Ionen Wolkens <ionen@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>