recover-broken-vdb: add --deep option (don't skip other paths)
We skip paths not matching *bin* and *libexec* (AND with a filename
that doesn't look like it ends in .so) given they're unlikely to be
critical or needed to bring up the system.

Add a --deep option to allow optionally considering these which
significantly increases runtime because we have to call file(1)
far more often.

(Noticed after hyphoon on IRC reported the scan tool showing
an issue with e.g. dev-tcltk/itk but the recovery tool
skipping it, because it doesn't install any files to
"important enough paths", although that's not the only/real
issue there, it made me realise.)

Signed-off-by: Sam James <sam@gentoo.org>